renamed:    src/panes/GamepadDebug.jsx -> src/components/GamepadDebug.jsx 	modified:   src/panes/DriveView.jsx
